Add pending_updates collector
This collector executes a command which is expected to return
pending updates for the system as <all>;<security>.

There is unfortunately no standardized way to pending updates for a
given distribution. Not even across deb based systems. Instead of
implementing a specific collector for deb based systems by jungling
around apt-get / apt-cache commands, this just defers the collection of
the actual updates to a script etc which returns <all>;<security>.

For ubuntu such script is provided by update-notifier-common which is
set as default here.