Add some basic battery statistics the MegaCLI collector.
While replacing some of our existing Icinga-based alerts with ones based
on Prometheus, we noticed that the MegaCLI exporter doesn't provide
insight in battery-related metrics.

This change adds metrics for some of the basic battery parameters
(voltage, current, temperature). Here at work we're also interested in
having a metric for whether periodic learning is required, as we'd
like to perform that manually, rather than letting the card do that
automatically.

Also hook up the existing adapter statistics to the end-to-end test.