Add a collector for Ksplice's Uptrack.
As we use Ksplice on all of our servers to apply kernel security patches
at run-time, we're interested in keeping an eye on its state to see
whether it properly applies all patches.

This collector mimicks the probing done by the commonly used
'check_uptrack_local' Icinga plugin. That plugin parses Ksplice's
'upgrade_plan' YAML file, extracting the number of changes that can be
applied, together with the timestamp at which the YAML file was
generated.

Due to the small size of this plugin, I've decided to not write any unit
tests. The end-to-end test already provides perfect coverage of the
plugin's functionality. This will obviously be reconsidered if
additional features are added.