Re-introduce human-readable chip name
The chip label generation has been changed in #334 to prefer the
unique device path (e.g. the location on the PCI bus) due to #333.

Here, a new label, chipName, is introduced which, again, carries
the human-readable sensor name (e.g. coretemp). It is used in
addition to the existing labels.

This allows to mitigate the downsides of the solution to #333
(namely that the device path may not be stable across kernels and
reboots) for cases where it does not matter that multiple devices
may have the same human-readable name (e.g. aggregation or where
at most one device of a type is present).