Expose systemd internal metrics
First of all, state changes of individual systemd units are now exposed.

Things like: When was the unit started, when was it stopped, how many
times did it restart before crash? How long until a timer activates.

Second of all, behind an optional flag --collector.systemd.accounting
additional metrics are defined, which expose systemd's view of CGroup
statistics of a unit. These metrics are also exposed in cAdvisor, and to
avoid duplicate metrics, this metric is explicitly opt in.