Add Linux inotify status support (#866)
If a system's inotify resources are exhausted, this can lead to
breakage all over the place (such as in systemd), which is hard to
spot otherwise.

This commit adds support for monitoring the inotify status.
More specifically, it tries initializing inotify
(node_inotify_init_success) and creating a watcher
for the root file system (node_inotify_add_watch_success).

It would be even nicer if the actual usage counters could be monitored.
However, the Linux kernel currently does not seem to expose those.

End-to-end tests for this collector are disabled as there does not seem
to be a way to mock the results.
Manual tests can be run by causing actual failures via:
  - sysctl fs.inotify.max_user_instances=0 or
  - sysctl fs.inotify.max_user_watches=0

Resource leakage tests have been performed by running apachebench
against /metrics and confirming that inotify does not break on its own.

Signed-off-by: Christian Hoffmann <mail@hoffmann-christian.info>