Add an option to override the instance label
Not all system report their instance as ip or hostname, it is still very
helpful to be able to correlate the node_exporter information and the
specific system.

It's true that for some cases it's possible to solve the issue with
label replacement at the prometheus server side.

But it's not always easy, especially when the prometheus query by IP and
the system report their own predefined name.

This patch adds the ability to configure the instance label.
It extend the labels creation to accept an array of additional labels
instead of a single one.

The change is limit to the text export, as the support for the protobuf
is being dropped.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>