Add an option to override the instance label
Not all system report their instance as ip or hostname, it is still very
helpful to be able to correlate the node_exporter information and the
specific system.

It's true that for some cases it's possible to solve the issue with
label replacement at the prometheus server side.

But it's not always easy, especially when the prometheus query by IP and
the system report their own predefined name.

This patch adds the ability to configure the instance label.
It wrap the metrics creation in a helper function. This function would
add the instance label if it was configured.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>