mixins: use 5m rate range for better interoperability
1m range limits usage of this mixin to environments where
scrape_interval is set to 15s or less. By extending this to 5m mixin
won't loose much value in regards to data, but it will be also more
usable in environments with less frequents scrapes. For example scrape
every 30s is fairly common practice, but data gathered from such
frequency cannot be used with this mixin.

Signed-off-by: paulfantom <pawel@krupa.net.pl>