Add tainted collector
Collect individual tainted flags as reported by the Linux kernel in
/proc/sys/kernel/tainted. For more information, see
https://www.kernel.org/doc/html/latest/admin-guide/tainted-kernels.html.

Tainted flags report the state of the system.  This state is primarily
intended for identifying unsupported conditions related to the Linux
kernel, ie inclusion in stack traces attached to bug reports to
identify that the issue occurred while the system was in an
unsupported state.  Some of the flags are useful to system operators
because of the undesired state they represent that needs remediation,
eg MCE, Oops, BUG.  Collecting these flags allow system operators to
be alerted to flag changes and keep metrics on flag values.

Signed-off-by: Corey Wright <cwright@digitalocean.com>