Change timex sync_status to follow maximum error
Some NTP implementations (e.g. chrony) can be configured to not clear
the "unsynchronized" bit from the system clock status in order to
prevent kernel updating the real-time clock, which causes the timex
sync_status to be always 0, even if the system clock is synchronized.

The kernel tracks a maximum error of the clock, which is incremented at
a rate of 500 ppm. When it gets to 16 seconds, the "unsynchronized" bit
is set automatically.

Ignore the timex status and set sync_status to 1 if the maximum error
is smaller than 16 seconds.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>