text_collectors: Use mktemp in scripts for simplicity
mktemp is more idiomatic and replaces the multiple steps with a
single statement. We must set all read on the file in order
for the other container to see it (mktemp does not use the default
filemask of the volume). An alternate approach may be used in the
future that involves giving the containers the same selinux context.