Replace load average with PSI metric
The load average metric is misleading as a representation of CPU
saturation. Normal CPU utilization is a better real representation of
saturation.

On newer Linux, there is a new Pressure Stall Information[0] metric that
better represents CPU over saturation. This is also useful as it can
make single-core saturation more visible.

[0]: https://www.kernel.org/doc/html/latest/accounting/psi.html

Signed-off-by: Ben Kochie <superq@gmail.com>