Attempt to sign the node exporter darwin build
This should hopefully fix the SIGKILL issue on OSX machines.
e.g. in: https://github.com/prometheus/node_exporter/issues/2539

Signed-off-by: Alper Polat <gitperr@gmail.com>

Change the docker flags to correct ones

Signed-off-by: Alper Polat <gitperr@gmail.com>

Fix errors in running the rcodesign from golang-builder

Signed-off-by: Alper Polat <gitperr@gmail.com>

Use pwd instead

Readlink does not work to get the proper path, pwd might do it.
As promu seems to be copying the binaries based on working directory.

Signed-off-by: Alper Polat <gitperr@gmail.com>

Try to run at the same job to see if it helps

So far I am unable to find the binary's location with
either pwd or readlink. I'm suspecting that the binary is
not on this specific host that is running the rcodesign.

Signed-off-by: Alper Polat <gitperr@gmail.com>

Try to debug what files are in the current working directory

Signed-off-by: Alper Polat <gitperr@gmail.com>

Print working directory as well

Signed-off-by: Alper Polat <gitperr@gmail.com>

Add quote wrapping

Signed-off-by: Alper Polat <gitperr@gmail.com>

Try to debug more

Signed-off-by: Alper Polat <gitperr@gmail.com>

Nothing seems to be in .build directory here

Signed-off-by: Alper Polat <gitperr@gmail.com>

Remove some of debug commands

Seems like the build does not get produced because of the
CircleCI node index that gets passed into `--parallelism-thread`.
Signed-off-by: Alper Polat <gitperr@gmail.com>

Add a separate sign stage for code signing

Separate stage might be useful so that we have all of
the builds that end up in `.build` here, and sign the one(s)
that we want. First one being implemented here is darwin-arm64.

Signed-off-by: Alper Polat <gitperr@gmail.com>

Run only if darwin-arm64 was built

Earlier I tried to add a separate stage for signing,
but seems like that was a bad idea because the pipeline
file has to exist in `master` for that so we can run
the tests properly. Checking with if might be one of the
simpler and better ideas...

Signed-off-by: Alper Polat <gitperr@gmail.com>

Add forgotten quote

Fixing basic syntax error

Signed-off-by: Alper Polat <gitperr@gmail.com>