feat!: new meminfo_procfs collector, deprecate meminfo collector
Part of #2957

Adds a new collector named `meminfo_procfs` that exposes memory metrics
in a format that attemps to be more inline with upstream conventions --
memory metrics are now exposed under a single metric named
`node_memory_bytes` that has a single label called `field`,
corresponding to the name of the field in `/proc/meminfo` that value
represents. Label values for the `field` label are named according to
the struct field in the procfs.Meminfo struct, and the values always use
the byte-normalized counterpart fields in the procfs.Meminfo struct,
resulting in a transition such as the following:
`node_memory_Active_anon_bytes -> node_memory_bytes{field="ActiveAnon"}`

Notes:
currently linux only, as that is the focus of the procfs lib. once
consensus has been reached here on new metric name/labels/format, I can
expand coverage for darwin/openbsd/netbsd and forward-port the existing
meminfo collector for those platforms to use the updated format.

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>